refactor: move buyDomainButton to subcomponent
