+++ src/features/buy-domain/BuyDomainButton/BuyDomainButton.module.scss
@@ -1,0 +1,8 @@
+.BuyButton {
+	&.Alternative {
+		margin-top: 1rem;
+		display: flex;
+		justify-content: center;
+		align-items: center;
+	}
+}
