+++ src/features/buy-domain/BuyDomainButton/index.ts
@@ -1,0 +1,1 @@
+export * from './BuyDomainButton';
