+++ src/features/buy-domain/BuyDomainModal/Elements/Mint.tsx
@@ -5,7 +5,8 @@ import { useBuyDomain, useWeb3, useZnsSdk } from '../../../../lib/hooks';
 import { DEFAULT_NETWORK_PROTOCAL } from '../../../../lib/constants/network';
 
 import { DefaultDomainMedia } from '.';
-import { DomainSearch } from '../../../domain-search/DomainSearch';
+
+import { DomainSearch } from '../../../../features/buy-domain/DomainSearch';
 import { QuestionButton } from '../../../ui/QuestionButton';
 import { Wizard, Tooltip } from '@zero-tech/zui/components';
