+++ src/features/buy-domain/DomainSearch/index.ts
