+++ src/features/buy-domain/index.ts
@@ -1,2 +1,4 @@
+export * from './BuyDomainButton';
 export * from './BuyDomainModal';
+export * from './DomainSearch';
 export * from './Icons';
