+++ src/pages/BuyDomain/BuyDomain.tsx
@@ -4,11 +4,11 @@ import { useState } from 'react';
 import { useWeb3 } from '../../lib/hooks';
 
 import { ConnectWallet } from '../../features/ui/ConnectWallet';
+import { IconZero } from '../../features/buy-domain/Icons';
+import { DomainSearch } from '../../features/buy-domain/DomainSearch';
+import { BuyDomainModal } from '../../features/buy-domain/BuyDomainModal';
 
 import styles from './BuyDomain.module.scss';
-import { IconZero } from '../../features/buy-domain/Icons';
-import { DomainSearch } from '../../features/domain-search';
-import { BuyDomainModal } from '../../features/buy-domain/BuyDomainModal';
 
 export const BuyDomain: FC = () => {
 	const { account } = useWeb3();
